New function for updating a config section
It is now possible to add new subsections and/or keys to an existing
section. To achieve this, the respective _config_section_update function
must be used to set the new default values.

Returning True in _config_section_update, will mean that the specific
section has been updated and it will be overwritten with the new values.

Change-Id: I155096814a00ddf9cccc3444f95e76b6539b41ed
Signed-off-by: Eduardo Lima (Etrunko) <eduardo.lima@intel.com>