Replace default theme font with SourceSansPro
DroidSans has missing glyphs that are used in the keyboard and has not
been updated for quite some time. It is being replaced in favor of
SourceSansPro, another open source font which provides all glyphs we
currently use.

The font is developed openly in github:
  - https://github.com/adobe-fonts/source-sans-proand

And distributed under the Open Font License:
  - http://www.adobe.com/type/browser/legal/pdfs/OpenFontLicense.pdf

Change-Id: I6e3cab82198ba860fe209d81982576b9bc0ecb8d
Signed-off-by: Eduardo Lima (Etrunko) <eduardo.lima@intel.com>